Extract slide component hide on sliders on mobile
